errors: rewrite to be backwards compatible with go stdlib
The errors.Error type is now a struct, not an interface, to prevent
extended methods from being lost when passed as an argument to a
function expecting type error. Hence, the top-level functions in the
errors package now expect type error, and detect if the error type is
extended to type errors.Error through the use of type switches.